Update ItemController Actions to Consider User
Small fix

+++ WishList/Controllers/ItemController.cs
@@ -45,8 +45,17 @@ namespace WishList.Controllers
 
         public IActionResult Delete(int id)
         {
+
+            var user = _userManager.GetUserAsync(HttpContext.User).Result;
+
+           
             var item = _context.Items.FirstOrDefault(e => e.Id == id);
-            _context.Items.Remove(item);
+
+            if (user.Id != item.User.Id)
+            {
+                return Unauthorized();
+            }
+                _context.Items.Remove(item);
             _context.SaveChanges();
             return RedirectToAction("Index");
         }
